Fixing tests in test_llm_manager and moving there the test on test_openai_providers.
